+++ pkg/config/config.go
@@ -15,7 +15,7 @@ import (
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 	"github.com/grafana/agent/pkg/config/features"
-	"github.com/grafana/agent/pkg/integrations/v2/app_o11y_receiver"
+	"github.com/grafana/agent/pkg/integrations/v2/app_agent_receiver"
 	"github.com/grafana/agent/pkg/logs"
 	"github.com/grafana/agent/pkg/metrics"
 	"github.com/grafana/agent/pkg/server"
@@ -186,20 +186,20 @@ func (c *Config) Validate(fs *flag.FlagSet) error {
 
 	c.Metrics.ServiceConfig.APIEnableGetConfiguration = c.EnableConfigEndpoints
 
-	// find every traces instance that is to be used by app o11y receiver
+	// find every traces instance that is to be used by app agent receiver
 	// and configure it to include a PushReceiver to be able to push traces internally
 	// this cannot happen in unmarshall stage since integrations are not yet unmarshalled
 	if c.Integrations.configV2 != nil {
 		for _, conf := range c.Integrations.configV2.Configs {
-			if conf.Name() != app_o11y_receiver.IntegrationName {
+			if conf.Name() != app_agent_receiver.IntegrationName {
 				continue
 			}
-			o11yConf := conf.(*app_o11y_receiver.Config)
-			if o11yConf.ReceiverConfig.TracesInstance == "" {
+			appAgentReceiverConf := conf.(*app_agent_receiver.Config)
+			if appAgentReceiverConf.ReceiverConfig.TracesInstance == "" {
 				continue
 			}
 			for idx, instanceConf := range c.Traces.Configs {
-				if instanceConf.Name == o11yConf.ReceiverConfig.TracesInstance {
+				if instanceConf.Name == appAgentReceiverConf.ReceiverConfig.TracesInstance {
 					c.Traces.Configs[idx].PushReceiver = true
 				}
 			}