+++ pkg/integrations/v2/app_agent_receiver/utils.go
@@ -1,4 +1,4 @@
-package app_o11y_receiver
+package app_agent_receiver
 
 import (
 	"fmt"
@@ -75,6 +75,7 @@ func KeyValToInterfaceMap(kv *KeyVal) map[string]interface{} {
 
 // URLMatchesOrigins returns true if URL matches at least one of origin prefix. Wildcard '*' and '?' supported
 func urlMatchesOrigins(URL string, origins []string) bool {
+	fmt.Println("matches", URL, origins)
 	for _, origin := range origins {
 		if origin == "*" || wildcard.Match(origin+"*", URL) {
 			return true